Fix overflow issue in tensor shape/stride

+++ pytorch_helpers/synapse_helpers/habana_tensor.h
@@ -118,7 +118,7 @@ class tensor final {
 
   class shape_t {
    public:
-    using dimension_size_t = unsigned;
+    using dimension_size_t = int64_t;
     struct dimension_count_t {
       explicit dimension_count_t(unsigned arg = 0) : value{arg} {}
       bool operator==(const dimension_count_t& rhs) const {
